requirements.txt
Adafruit-Blinka==8.64.0
adafruit-circuitpython-busdevice==5.2.13
adafruit-circuitpython-connectionmanager==3.1.5
adafruit-circuitpython-pixelbuf==2.0.9
adafruit-circuitpython-requests==4.1.13
adafruit-circuitpython-seesaw==1.16.7
adafruit-circuitpython-typing==1.12.1
Adafruit-PlatformDetect==3.82.0
Adafruit-PureIO==1.1.11
binho-host-adapter==0.1.6
gphoto2==2.6.2
numpy==2.2.6
opencv-python==4.12.0.88
pillow==11.3.0
pyftdi==0.57.1
pyserial==3.5
pyusb==1.3.1
RPi.GPIO==0.7.1
spidev==3.7
sysv-ipc==1.1.0
typing_extensions==4.14.1
